Drastically simplified character escaping
I took another look at the specification for Turtle
(https://www.w3.org/TR/turtle/#h3_literals) and found that only newline,
backslashes and quotes really needed to be escaped. What N3 does is
probably not incorrect, but just seems to complicated for now. So, this
is now my own implementation of an escape mechanism. No copies this
time, just inspiration from the spec and what also appears to be done in
https://github.com/linkeddata/rdflib.js. If this still looks like
copying, this is just basically how anyone could write it based on the
spec.

Oh, yes, I also removed escaping from the IRI of a NamedNode. It feels
like this library is not responsible for unsupported characters in IRIs.
The same could be said about what is inside a string literal, but for
that case, newlines and such *are allowed* if the string writer formats
it with the correct delimiters ("""). Our approach here is to escape
characters instead as a simplification.